requirements.txt
# =============================================================================
# Fantasy Basketball Optimizer - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Flask Core and Extensions
# -----------------------------------------------------------------------------
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
Werkzeug==3.0.1

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
SQLAlchemy==2.0.25
# Note: SQLite is built into Python, no extra package needed for development

# PostgreSQL adapter (uncomment for production deployment):
# psycopg2-binary==2.9.9
# Or install PostgreSQL and use: pip install psycopg2

# -----------------------------------------------------------------------------
# Environment & Configuration
# -----------------------------------------------------------------------------
python-dotenv==1.0.1

# -----------------------------------------------------------------------------
# ESPN Fantasy API
# -----------------------------------------------------------------------------
espn-api==0.45.1

# -----------------------------------------------------------------------------
# Data Science & Machine Learning
# -----------------------------------------------------------------------------
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.12.0               # Scientific computing (sklearn dependency)
joblib==1.3.2               # Model serialization/persistence

# -----------------------------------------------------------------------------
# Web Scraping (Basketball Reference)
# -----------------------------------------------------------------------------
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0                 # Fast HTML/XML parser
urllib3==2.1.0              # HTTP library

# -----------------------------------------------------------------------------
# Task Scheduling
# -----------------------------------------------------------------------------
APScheduler==3.10.4

# -----------------------------------------------------------------------------
# Security & Cryptography
# -----------------------------------------------------------------------------
cryptography==42.0.2        # Cookie encryption, secure operations
bcrypt==4.1.2               # Password hashing (Flask-Bcrypt dependency)

# -----------------------------------------------------------------------------
# Date/Time Utilities
# -----------------------------------------------------------------------------
python-dateutil==2.8.2
pytz==2024.1                # Timezone support

# -----------------------------------------------------------------------------
# Data Validation
# -----------------------------------------------------------------------------
email-validator==2.1.0      # Email validation

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest==8.0.0
pytest-cov==4.1.0           # Test coverage
pytest-flask==1.3.0         # Flask testing utilities
pytest-mock==3.12.0         # Mocking utilities
coverage==7.4.1             # Code coverage reporting

# -----------------------------------------------------------------------------
# Code Quality (Development)
# -----------------------------------------------------------------------------
# Uncomment these for development:
# black==24.1.1             # Code formatter
# flake8==7.0.0             # Linter
# isort==5.13.2             # Import sorter
# mypy==1.8.0               # Type checker

# -----------------------------------------------------------------------------
# Optional: Rate Limiting
# -----------------------------------------------------------------------------
# Uncomment for API rate limiting:
# Flask-Limiter==3.5.1
# limits==3.7.0

# -----------------------------------------------------------------------------
# Optional: Caching
# -----------------------------------------------------------------------------
# Uncomment for Redis caching (production):
# redis==5.0.1
# Flask-Caching==2.1.0

# -----------------------------------------------------------------------------
# Optional: Database Migrations
# -----------------------------------------------------------------------------
# Uncomment when ready for production migrations:
# Flask-Migrate==4.0.5
# alembic==1.13.1

# -----------------------------------------------------------------------------
# Optional: API Documentation
# -----------------------------------------------------------------------------
# Uncomment for auto-generated API docs:
# flask-restx==1.3.0
# flasgger==0.9.7.1

# -----------------------------------------------------------------------------
# Optional: Anthropic API (AI Features)
# -----------------------------------------------------------------------------
# Uncomment for advanced AI trade suggestions:
# anthropic==0.18.1

# -----------------------------------------------------------------------------
# Production Server (for deployment)
# -----------------------------------------------------------------------------
# Uncomment for production:
# gunicorn==21.2.0
# gevent==24.2.1
